+++ lib/xdr/concerns/reads_bytes.rb
@@ -5,4 +5,12 @@ module XDR::Concerns::ReadsBytes
       raise EOFError if bytes.nil? || bytes.length != length
     end
   end
+
+  def read_zeros(io, length)
+    read_bytes(io, length).each_byte do |byte|
+      raise XDR::ReadError unless byte == 0
+    end
+
+    nil
+  end
 end