+++ lib/xdr/opaque.rb
@@ -12,7 +12,7 @@ class XDR::Opaque
   def read(io)
     # read and return @length bytes
     # throw away @padding bytes
-    read_bytes(io, @length).tap{ read_bytes(io, @padding) }
+    read_bytes(io, @length).tap{ read_zeros(io, @padding) }
   end
 
   def write(val,io)