+++ spec/lib/xdr/opaque_spec.rb
@@ -9,6 +9,10 @@ describe XDR::Opaque, "#read" do
     expect(read("\x00\x01\x00\x00")).to eq("\x00\x01\x00")
   end
 
+  it "raises a ReadError when the padding isn't zeros" do
+    expect{ read "\x00\x00\x01\x01" }.to raise_error(XDR::ReadError)
+  end
+
   def read(str)
     io = StringIO.new(str)
     subject.read(io)